Update Flight + flight generation and init support
Removed entity Flight and added FlightTemplate and FlightInstance entities in its place and corresponding repositories. Added Flights.json and support for init of flight data such as FlightDBInit component, added initial support for FlightJSON generation with a FlightTemplateGenerationRequest DTO, added FlightGenerationProperties for SSOT of flight data such as prices, price multipliers and orbit times. Added FlightGenerationService to to support and delegate all of the above. Added AdminController and endpoint for generating FlightJSONs